registracia opravy

+++ SportApp/Core/src/main/java/com/so/core/services/RegistrationService.java
@@ -6,6 +6,7 @@
 package com.so.core.services;
 
 import com.so.core.controller.converter.RegistrationConverter;
+import com.so.core.controller.dto.ResourceDto;
 import com.so.core.controller.dto.registration.RegistrationPlayerDto;
 import com.so.core.controller.dto.registration.RegistrationTeamDto;
 import com.so.core.services.document.DocumentService;
@@ -16,7 +17,9 @@ import com.so.dal.core.repository.registration.RegistrationPlayerRepository;
 import com.so.dal.core.repository.registration.RegistrationTeamRepository;
 import java.io.IOException;
 import java.security.InvalidParameterException;
+import java.util.ArrayList;
 import java.util.Date;
+import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -29,25 +32,73 @@ import org.springframework.transaction.annotation.Transactional;
  */
 @Service
 public class RegistrationService {
-
+    
     private final static Logger LOG = LoggerFactory.getLogger(RegistrationService.class);
-
+    
     @Autowired
     private RegistrationTeamRepository regTeamRepo;
-
+    
     @Autowired
     private RegistrationPlayerRepository regPlayerRepo;
-
+    
     @Autowired
     private DocumentService documentService;
-
+    
     @Autowired
     private RegistrationConverter converter;
-
+    
+    @Transactional
+    public List<RegistrationTeamDto> getAllTeams() {
+       LOG.info("getAllTeams()");
+        List<RegistrationTeamDto> lDto = new ArrayList<>();
+        
+        List<RegistrationTeam> lEntity = regTeamRepo.findAll();
+        
+        for (RegistrationTeam t : lEntity) {
+            lDto.add(converter.regTeamEntityToDto(t, false));
+        }
+        return lDto;
+    }
+    
+  
+    @Transactional
+    public RegistrationTeamDto getTeam(Integer id) {
+        LOG.info("getTeam({})", id);
+   
+        if(id==null){
+            LOG.error("nevyplneny povinny atribut id={}",id);
+            throw new InvalidParameterException("nevyplneny povinny atribut id");
+        }
+        RegistrationTeam team = regTeamRepo.findOne(id);
+        
+        if (team == null) {
+            LOG.info("nenajdeny ziadny tim podla id={}", id);
+            return new RegistrationTeamDto();
+        }
+        return converter.regTeamEntityToDto(team, true);
+    }
+    
+      @Transactional
+    public RegistrationPlayerDto getPlayer(Integer id){
+              LOG.info("getPlayer({})", id);
+   
+        if(id==null){
+            LOG.error("nevyplneny povinny atribut id={}",id);
+            throw new InvalidParameterException("nevyplneny povinny atribut id");
+        }
+        
+        RegistrationPlayer p = regPlayerRepo.findOne(id);
+        if(p==null){
+            LOG.info("nenajdeny ziadny regPlayer pre id={}",id);
+            return new RegistrationPlayerDto();
+        }
+        return converter.regPlayerEntityToDto(p);
+    }
+    
     @Transactional
     public RegistrationTeamDto registrationTeam(RegistrationTeamDto teamDto) throws IOException {
         LOG.info("registrationTeam(teamDto)");
-
+        
         if (teamDto.getColor() == null || teamDto.getName() == null || teamDto.getShortName() == null
                 || teamDto.getZnak() == null || teamDto.getRegistrationPlayers() == null || teamDto.getSeasonTournamentId() == null) {
             LOG.error("nevyplneny povinny atribut color={} name={} shortName={} znak={} players={} seasonTournamentId= {}", teamDto.getColor(),
@@ -57,31 +108,25 @@ public class RegistrationService {
         Resource znak = documentService.createFile(teamDto.getZnak().getData(), teamDto.getZnak().getMimeType());
         teamDto.getZnak().setId(znak.getId());
         teamDto.setCreatedTime(new Date());
-
+        
         RegistrationTeam team = converter.regTeamDtoToEntity(teamDto, false);
         RegistrationTeam savedTeam = regTeamRepo.saveAndFlush(team);
-
+        
         if (savedTeam == null) {
             LOG.error(" do databazy sa nepodarilo ulozit team={} ", team);
             throw new IllegalStateException("do databazy sa neulozil novo vytvoreny team, koncim spracovanie");
         }
-
+        
         for (RegistrationPlayerDto p : teamDto.getRegistrationPlayers()) {
-            try {
-                savedTeam.getRegistrationPlayers().add(registrationPlayer(p, savedTeam));
-            } catch (IllegalStateException e) {
-                //catch aby ak jeden hrac zakape dalsi sa ulozili .... ake najlepsie riesenie?
-                // bud prejde vsetko alebo nic? alebo ulozime dobrych hracov??
-
-            }
+            savedTeam.getRegistrationPlayers().add(registrationPlayer(p, savedTeam));
         }
-        return converter.regTeamEntityToDto(savedTeam); // vraciam z dbs aby mal uz aj odkaz na playerov
+        return converter.regTeamEntityToDto(savedTeam, true); // vraciam z dbs aby mal uz aj odkaz na playerov
     }
-
+    
     @Transactional
-    private RegistrationPlayer registrationPlayer(RegistrationPlayerDto playerDto, RegistrationTeam team) {
+    private RegistrationPlayer registrationPlayer(RegistrationPlayerDto playerDto, RegistrationTeam team) throws IOException {
         LOG.info("registrationPlayer(playerDto, team)-> playerDto={} teamID={}", playerDto, team.getId());
-
+        
         if (playerDto.getIsProfessional() == null || playerDto.getIsStudent() == null || playerDto.getName() == null
                 || playerDto.getSurname() == null || playerDto.getNumber() == null || playerDto.getSex() == null) {
             LOG.error("nevyplnene povinne parametre isProfessional={} isStudent={} name={} surName={} Number={} sex={}",
@@ -89,8 +134,15 @@ public class RegistrationService {
                     playerDto.getNumber(), playerDto.getSex());
             throw new InvalidParameterException("nevyplnene povinne parametre");
         }
-
+        
+        Resource r = null;
+        
+        if (playerDto.getPhoto() != null) {
+            r = documentService.createFile(playerDto.getPhoto().getData(), playerDto.getPhoto().getMimeType());
+        }
         RegistrationPlayer player = converter.regPlayerDtoToEntity(playerDto);
+        
+        player.setPhoto(r);
         player.setRegistrationTeam(team);
         RegistrationPlayer savedPlayer = regPlayerRepo.saveAndFlush(player);
         if (savedPlayer == null) {
@@ -99,12 +151,12 @@ public class RegistrationService {
         }
         return savedPlayer;
     }
-
+    
     @Transactional
     public void deleteRegistrationTeam(Integer id) {
-
+        
         RegistrationTeam team = regTeamRepo.findOne(id);
-
+        
         if (team == null) {
             throw new InvalidParameterException("nenajdeny tim s id:" + id);
         }
@@ -114,53 +166,58 @@ public class RegistrationService {
             deleteRegistrationPlayer(p.getId());
         }
         regTeamRepo.delete(id);
-
-        if (team.getResource().getCompetitorTeams().isEmpty()) {
+        
+        if (team.getResource().getCompetitorTeams().isEmpty()
+                && team.getResource().getRegistrationTeams().size() == 1) {
             documentService.deleteFile(team.getResource());
         }
-
+        
     }
-
+    
     @Transactional
     public void deleteRegistrationPlayer(Integer id) {
-
+        
         RegistrationPlayer p = regPlayerRepo.findOne(id);
-
+        
         if (p == null) {
             throw new InvalidParameterException("nenajdeny hrac s :" + id);
         }
-
+        
         regPlayerRepo.delete(id);
 
         //TODO doplnit vymazenie resourcov ak sa doplni do db
         //   if(p.g)
+        if (p.getPhoto().getCompetitorTeamPlayers().isEmpty()
+                && p.getPhoto().getRegistrationPlayers().size() == 1) {
+            documentService.deleteFile(p.getPhoto());
+        }
     }
-
+    
     @Transactional
     public RegistrationTeamDto editTeam(RegistrationTeamDto team) {
-
+        
         RegistrationTeam entity = converter.regTeamDtoToEntity(team, false);
         RegistrationTeam savedTeam = regTeamRepo.save(entity);
-
+        
         if (savedTeam == null) {
             LOG.error("nepodarilo sa ulozit aktualizovany tim do db timId={}", team.getId());
             throw new IllegalStateException("nepodarilo sa ulozit aktualizovany tim do db");
         }
-
-        return converter.regTeamEntityToDto(savedTeam);
+        
+        return converter.regTeamEntityToDto(savedTeam, false);
     }
-
+    
     @Transactional
     public RegistrationPlayerDto editPlayer(RegistrationPlayerDto player) {
-
+        
         RegistrationPlayer entity = converter.regPlayerDtoToEntity(player);
         RegistrationPlayer savedPlayer = regPlayerRepo.saveAndFlush(entity);
-
+        
         if (savedPlayer == null) {
             LOG.error("nepodarilo sa ulozit aktualizovaneho hraca do db playerId={}", player.getId());
             throw new IllegalStateException("nepodarilo sa ulozit aktualizovaneho hraca do db");
         }
-
+        
         return converter.regPlayerEntityToDto(savedPlayer);
     }
 }
